gdbstub: SendTrap(): If thread check fails, return no-op
The original check is hard to explain at a first glance, so I
made its intent clearer by negating the condition--returning
a complete no-op if not met--and by replacing 'halt_flag'
with 'step_flag', as both flags work as a pair.
This check is relevant whenever GDB has issued a Step() command to our stub.

We don't want a failed check to set any of these flags:

** (halt_flag, step_flag) to (true, false).
This flag combination leaves the CPU halted, without allowing time to advance for ThreadManager to switch threads. Granted, both Step() and RunLoop() had enforced this state to begin with. Thus, in this case, the statement 'halt_flag = true' from SendTrap has no side effect.

** send_trap to false. It stops future calls to SendTrap from reporting anything to gdb ever again. We would rather leave a Stop Reply pending than forgotten.

Unfortunately, these changes alone don't suffice to avoid
the softlock.  System::RunLoop needs additional changes.
That's because it never calls SendTrap again if the CPU
enters a Halted state and remains so.